Move metadata tests to metadata.rs
